something is going weird
In the piece editor section, when I pass the field variable as this.name,
the actual object passed is the name variable of the game table.
I tried a bunch of things for hours, and couldn't quite narrow it down.
Also, passing piece[i]['whatever'] is undefined. I think it's because 
i isn't being hardcoded into the dom object like I thought. Need a
workaround.